chatbot/requirements.txt
streamlit>=1.32.0
google-generativeai>=0.8.0
mysql-connector-python>=8.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
mcp[cli]>=1.0.0
qdrant-client>=1.7.0
httpx>=0.27.0
